refactor: poll and wait utils
