+++ src/components/modal/components/Chat/components/EscalateModal/steps/EscalateStepTwo.tsx
@@ -11,6 +11,7 @@ import { useWeb3React } from "@web3-react/core";
 import { useConfigContext } from "components/config/ConfigContext";
 import { BigNumber, BigNumberish, ethers, utils } from "ethers";
 import { Form, Formik, FormikProps, FormikState } from "formik";
+import { poll } from "lib/utils/promises";
 import {
   Dispatch,
   SetStateAction,
@@ -37,7 +38,6 @@ import { useCoreSDK } from "../../../../../../../lib/utils/useCoreSdk";
 import { useChatContext } from "../../../../../../../pages/chat/ChatProvider/ChatContext";
 import { ICON_KEYS } from "../../../../../../../pages/chat/components/conversation/const";
 import { MessageDataWithInfo } from "../../../../../../../pages/chat/types";
-import { poll } from "../../../../../../../pages/create-product/utils";
 import Collapse from "../../../../../../collapse/Collapse";
 import { Checkbox } from "../../../../../../form";
 import { FieldInput } from "../../../../../../form/Field.styles";
