+++ src/components/modal/components/RevokeProduct.tsx
@@ -1,6 +1,7 @@
 import { Provider, RevokeButton, subgraph } from "@bosonprotocol/react-kit";
 import * as Sentry from "@sentry/browser";
 import { useConfigContext } from "components/config/ConfigContext";
+import { poll } from "lib/utils/promises";
 import toast from "react-hot-toast";
 import styled from "styled-components";
 
@@ -9,7 +10,6 @@ import { useEthersSigner } from "../../../lib/utils/hooks/ethers/useEthersSigner
 import { useAddPendingTransaction } from "../../../lib/utils/hooks/transactions/usePendingTransactions";
 import { Exchange } from "../../../lib/utils/hooks/useExchanges";
 import { useCoreSDK } from "../../../lib/utils/useCoreSdk";
-import { poll } from "../../../pages/create-product/utils";
 import { Break } from "../../detail/Detail.style";
 import Price from "../../price/index";
 import { useConvertedPrice } from "../../price/useConvertedPrice";
