+++ src/pages/create-product/utils/index.tsx
@@ -36,25 +36,6 @@ import { isTruthy } from "lib/types/helpers";
 import { ChatInitializationStatus } from "lib/utils/hooks/chat/useChatStatus";
 import React from "react";
 
-export const poll = async function <T>(
-  fn: () => Promise<T>,
-  fnConditionToKeepPolling: (arg: T) => boolean,
-  ms: number
-): Promise<T> {
-  let result = await fn();
-  while (fnConditionToKeepPolling(result)) {
-    await wait(ms);
-    result = await fn();
-  }
-  return result;
-};
-
-export const wait = async (ms: number) => {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-};
-
 export type CreateProductSteps = {
   0: {
     ui: JSX.Element;
